Add ability to tweet found search results at searched handle.
Each search result is rendered with a twitter share button that will auto-populate with the correct URL and twitter handle of the person originally searched to generate the keywords. Before, no tweet button was available and users could not tweet. This commit adds the functionality to tweet at the correct handle, and with the correct link. With this commit we finish our MVP.