Embed tweets and replace tweet text link. Remove profile pictures from each tweet.
Embedding tweets allows for the ability to reply to the tweet, favorite it, or retweet it all while one our site. Before, a user could not reply to a tweet, or watch embedded content such as videos, or view photos. Users can now also click on the @ mentioned names within tweets.