Return tweet rendering to prior state.
When embedding tweets we made far too many requests to the Twitter API and quickly exceeded our limit. By returning to the previous scraping strategy we make fewer requests and grab more tweets with each request, making it easier to stay within our limit. We lose the ability to embed videos, and reply to tweets.