Add explanatory text for users new to the site.
When a search is made, explanatory text on how to use the Connections web as well as the search result feature is provided. If the user does not refresh the page the text will not show again so as to keep the user experience smooth.

Also, urls returned in the search results have additional explanatory links.