Refine the animation flow
I had to change when the one degree button fades in and out in order to avoid the possibility of a user breaking the animations by pressing the button when we don't want them to. To do that, however, I had to remove the hatch animation from the button because it was causing a visual glitch when trying to fade the button back in.